requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.23.2
httpx==0.24.1
pydantic-settings==0.9.1
transformers==4.34.0
torch==2.1.0
python-dotenv==1.0.0
